# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /yard/spots/{spotId}:
    patch:
      operationId: updateYardSpot
      tags: [yard-spots]
      summary: Update a spot (assign, release, reserve vehicle)
      parameters:
        - name: spotId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateYardSpotBody"
      responses:
        "200":
          description: Updated spot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YardSpot"

  # ── Vehicles ──────────────────────────────────────────────────────────────
  /yard/vehicles:
    get:
      operationId: listYardVehicles
      tags: [yard-vehicles]
      summary: List vehicle inventory
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [available, in_transit, pdi_pending, sold, all]
        - name: locationId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 15
      responses:
        "200":
          description: Vehicle list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YardVehicleList"
    post:
      operationId: createYardVehicle
      tags: [yard-vehicles]
      summary: Add a new vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateYardVehicleBody"
      responses:
        "201":
          description: Created vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YardVehicle"
  /yard/vehicles/{vehicleId}:
    get:
      operationId: getYardVehicle
      tags: [yard-vehicles]
      summary: Get vehicle detail
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YardVehicle"
    patch:
      operationId: updateYardVehicle
      tags: [yard-vehicles]
      summary: Update vehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateYardVehicleBody"
      responses:
        "200":
          description: Updated vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YardVehicle"

  # ── Inspections ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    YardSpot:
      type: object
      properties:
        id:
          type: integer
        zoneId:
          type: integer
        code:
          type: string
        status:
          type: string
          enum: [available, occupied, reserved, disabled]
        vehicleId:
          type: integer
          nullable: true
        vehicle:
          $ref: "#/components/schemas/YardVehicle"
          nullable: true
        reservedUntil:
          type: string
          nullable: true
        dimensions:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        spotType:
          type: string
          nullable: true
        timeInSpot:
          type: string
          nullable: true
      required: [id, zoneId, code, status]
    UpdateYardSpotBody:
      type: object
      properties:
        status:
          type: string
          enum: [available, occupied, reserved, disabled]
        vehicleId:
          type: integer
          nullable: true
        reservedUntil:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true

    # ── Vehicle ──────────────────────────────────────────────────────────────
    YardVehicle:
      type: object
      properties:
        id:
          type: integer
        vin:
          type: string
        stockNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
          nullable: true
        mileage:
          type: integer
          nullable: true
        condition:
          type: string
          enum: [new, used]
          nullable: true
        status:
          type: string
          enum: [available, in_transit, pdi_pending, sold]
        locationId:
          type: integer
          nullable: true
        locationName:
          type: string
          nullable: true
        spotId:
          type: integer
          nullable: true
        spotCode:
          type: string
          nullable: true
        zoneName:
          type: string
          nullable: true
        price:
          type: number
          nullable: true
        imageUrl:
          type: string
          nullable: true
        arrivedAt:
          type: string
          nullable: true
      required: [id, vin, stockNumber, make, model, year, status]
    YardVehicleList:
      type: object
      properties:
        vehicles:
          type: array
          items:
            $ref: "#/components/schemas/YardVehicle"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [vehicles, total, page, totalPages]
    CreateYardVehicleBody:
      type: object
      properties:
        vin:
          type: string
        stockNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        mileage:
          type: integer
        condition:
          type: string
          enum: [new, used]
        status:
          type: string
          enum: [available, in_transit, pdi_pending, sold]
        locationId:
          type: integer
        price:
          type: number
      required: [vin, stockNumber, make, model, year, status]
    UpdateYardVehicleBody:
      type: object
      properties:
        status:
          type: string
          enum: [available, in_transit, pdi_pending, sold]
        locationId:
          type: integer
          nullable: true
        spotId:
          type: integer
          nullable: true
        color:
          type: string
        price:
          type: number

    # ── Inspection ────────────────────────────────────────────────────────────
...