# lib/api-zod/src/generated/types/updateYardVehicleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateYardVehicleBodyStatus } from "./updateYardVehicleBodyStatus";

export interface UpdateYardVehicleBody {
  status?: UpdateYardVehicleBodyStatus;
  locationId?: number | null;
  spotId?: number | null;
  color?: string;
  price?: number;
}